:sparkles: Enable CMAKE_EXPERIMENTAL_EXPORT_PACKAGE_DEPENDENCIES
This change also removes much of the conditional code that was actually
completely broken previously. The previous code could not work because
compiler details are not available at the time the toolchains are
injected into the build.

Import std support comes with some annoyances that make it problematic
if used with a compiler that doesn't support it, or even those that do.

For now, this package will simply enable the safe flags that cause no
problems even if modules are not used.